test: improve svgo suite based on mutation testing results
