+++ src/svgo/__mocks__/svgo-v1-wrapper.ts
@@ -1,3 +1,9 @@
-import SVGOptimizer from "./common";
+const optimize = jest.fn()
+  .mockReturnValue({ data: "<svg></svg>" })
+  .mockName("SVGOptimizerV2.optimize");
+
+const SVGOptimizer = jest.fn()
+  .mockReturnValue({ optimize })
+  .mockName("SVGOptimizer");
 
 export default SVGOptimizer;
